requirements.txt
python-telegram-bot==21.0
requests==2.31.0
beautifulsoup4==4.12.2
fake-useragent==1.4.0
python-dotenv==1.0.0
scrapfly-sdk

# Web application
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
jinja2==3.1.3
python-multipart==0.0.6
httpx==0.26.0

# Development
jupyter
